Removed Default appList fetch from AppSelector
