+++ src/components/v2/headers/ChartHeader.component.tsx
@@ -3,7 +3,7 @@ import { NavLink, Link } from 'react-router-dom';
 import { URLS, Routes, AppListConstants } from '../../../config';
 import { BreadCrumb, useBreadcrumb } from '../../common';
 import ReactGA from 'react-ga';
-import AppSelector from '../../AppSelector';
+import { ChartSelector } from '../../AppSelector';
 import { useParams, useRouteMatch, useHistory, generatePath } from 'react-router';
 import { get } from '../../../services/api';
 import { handleUTCTime } from '../../common';
@@ -47,7 +47,7 @@ function ChartHeaderComponent() {
             alias: {
                 ':appId(\\d+)': {
                     component: (
-                        <AppSelector
+                        <ChartSelector
                             //@ts-ignore
                             api={getInstalledCharts}
                             primaryKey="appId"
